add some packages

+++ prepare.sh
@@ -15,14 +15,29 @@ export SSID
 export WPA_PASSPHRASE
 export HOSTNAME
 
-curl -o install-stage0.sh https://raw.githubusercontent.com/cwebster2/environment-installer/master/install-stage0-${INSTALLER}.sh
+echo "* Validating system was booted in UEFI mode"
+if [ ! -d "/sys/firmware/efi" ]; then
+  echo "ERROR: Please reboot in UEFI mode."
+  exit 1
+fi
 
+echo "* Setting system time"
+systemctl start systemd-timesyncd
+
+echo "* Getting installer scripts"
+curl -sLo install-stage0.sh https://raw.githubusercontent.com/cwebster2/environment-installer/master/install-stage0-${INSTALLER}.sh
 chmod 755 install-stage0.sh
 
-echo "Running stage0 prepare.  You will be prompted for zfs passphrase and user password"
+echo "* Running stage0 prepare.  You will be prompted for zfs passphrase and user password"
 
 ./install-stage0.sh prepare
 
+echo "* Cleaning up and rebooting the system"
 rm install-stage0.sh
+
+cat <<-EOF > "/home/${TARGET_USER}/.zshrc
+export INSTALLER=${INSTALLER}
+./install.sh
+EOF
 
 reboot
